+++ packages/hardhat-zksync-vyper/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @matterlabs/hardhat-zksync-vyper
 
+## [1.1.0](https://github.com/matter-labs/hardhat-zksync/compare/@matterlabs/hardhat-zksync-vyper-v1.0.8...@matterlabs/hardhat-zksync-vyper-v1.1.0) (2024-09-05)
+
+
+### Features
+
+* bump ethers, zksync-ethers, hardaht and other dependencies to newer versions ([#1111](https://github.com/matter-labs/hardhat-zksync/issues/1111)) ([a2d503a](https://github.com/matter-labs/hardhat-zksync/commit/a2d503abe3f504859651f22998046576eddf6579))
+
+
+### Bug Fixes
+
+* add gnu toolchain for linux as default from compiler version 1.5.4 ([#1359](https://github.com/matter-labs/hardhat-zksync/issues/1359)) ([4272477](https://github.com/matter-labs/hardhat-zksync/commit/4272477e9a132ba7f36c24a14a4ac15ea65f1c65))
+* update links to new doc site ([276740b](https://github.com/matter-labs/hardhat-zksync/commit/276740ba5abf8b5775e135b5653824d6456a7e4f))
+
 ## [1.0.8](https://github.com/matter-labs/hardhat-zksync/compare/@matterlabs/hardhat-zksync-vyper-v1.0.7...@matterlabs/hardhat-zksync-vyper-v1.0.8) (2024-02-26)